web/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
pandas>=2.0.0
openpyxl>=3.1.0
pyxlsb>=1.0.10
pyyaml>=6.0
python-pptx>=0.6.21
numpy>=1.24.0

